+++ src/main/java/goorm/ddok/cafe/repository/CafeRepository.java
@@ -4,6 +4,7 @@ import goorm.ddok.cafe.domain.Cafe;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,5 +21,5 @@ public interface CafeRepository extends JpaRepository<Cafe, Long> {
           and c.activityLatitude between :swLat and :neLat
           and c.activityLongitude between :swLng and :neLng
         """)
-    List<Cafe> findActiveWithinBounds(double swLat, double swLng, double neLat, double neLng);
+    List<Cafe> findActiveWithinBounds(BigDecimal swLat, BigDecimal swLng, BigDecimal neLat, BigDecimal neLng);
 }
